# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Resume Shortlisting System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  open: 'open',
  closed: 'closed',
  paused: 'paused',
} as const;

export interface Job {
  id: number;
  title: string;
  description: string;
  requiredSkills: string[];
  experienceYears: number;
  status: JobStatus;
  candidateCount: number;
  createdAt: string;
}

export type JobInputStatus = typeof JobInputStatus[keyof typeof JobInputStatus];


export const JobInputStatus = {
  open: 'open',
  closed: 'closed',
  paused: 'paused',
} as const;

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requiredSkills: string[];
  experienceYears: number;
  status?: JobInputStatus;
}

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  open: 'open',
  closed: 'closed',
  paused: 'paused',
} as const;

export interface JobUpdate {
  title?: string;
  description?: string;
  requiredSkills?: string[];
  experienceYears?: number;
  status?: JobUpdateStatus;
}

export type CandidateStatus = typeof CandidateStatus[keyof typeof CandidateStatus];


export const CandidateStatus = {
  pending: 'pending',
  analyzed: 'analyzed',
  shortlisted: 'shortlisted',
  rejected: 'rejected',
} as const;

export interface Candidate {
  id: number;
  name: string;
  email: string;
  resumeText: string;
  jobId: number;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  status: CandidateStatus;
  verifiedProofOfWork: boolean;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  extractedSkills?: string[];
  createdAt: string;
}

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  resumeText: string;
  jobId: number;
  githubUrl?: string;
  portfolioUrl?: string;
  location?: string;
  university?: string;
  yearsExperience?: number;
}

export type CandidateUpdateStatus = typeof CandidateUpdateStatus[keyof typeof CandidateUpdateStatus];


export const CandidateUpdateStatus = {
  pending: 'pending',
  analyzed: 'analyzed',
  shortlisted: 'shortlisted',
  rejected: 'rejected',
} as const;

export interface CandidateUpdate {
  name?: string;
  email?: string;
  status?: CandidateUpdateStatus;
  githubUrl?: string;
  portfolioUrl?: string;
}

export type CandidateBlindStatus = typeof CandidateBlindStatus[keyof typeof CandidateBlindStatus];


export const CandidateBlindStatus = {
  pending: 'pending',
  analyzed: 'analyzed',
  shortlisted: 'shortlisted',
  rejected: 'rejected',
} as const;

export interface CandidateBlind {
  id: number;
  resumeText: string;
  jobId: number;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  status: CandidateBlindStatus;
  verifiedProofOfWork: boolean;
  extractedSkills: string[];
}

export interface AnalyzeRequest {
  jobId: number;
}

export type AnalysisResultRecommendation = typeof AnalysisResultRecommendation[keyof typeof AnalysisResultRecommendation];


export const AnalysisResultRecommendation = {
  shortlist: 'shortlist',
  reject: 'reject',
  hold: 'hold',
} as const;

export interface AnalysisResult {
  candidateId: number;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  matchedSkills: string[];
  missingSkills: string[];
  summary: string;
  recommendation: AnalysisResultRecommendation;
}

export interface CandidateFeedback {
  candidateId: number;
  emailSubject: string;
  emailBody: string;
  improvementAreas: string[];
}

export interface DashboardSummary {
  totalJobs: number;
  totalCandidates: number;
  shortlisted: number;
  rejected: number;
  pending: number;
  averageScore: number;
  /** @nullable */
  topScoringJob: string | null;
}

export interface CandidateRank {
  candidateId: number;
  candidateName: string;
  jobTitle: string;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  status: string;
}

export type CandidateSummaryVerdict = typeof CandidateSummaryVerdict[keyof typeof CandidateSummaryVerdict];


export const CandidateSummaryVerdict = {
  shortlist: 'shortlist',
  hold: 'hold',
  reject: 'reject',
} as const;

export interface CandidateSummary {
  candidateId: number;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  verdict: CandidateSummaryVerdict;
  pros: string[];
  cons: string[];
  highlights: string[];
  riskFlags: string[];
  skillsFound: string[];
  skillsMissing: string[];
  verifiedProofOfWork: boolean;
}

export interface SkillGap {
  skill: string;
  jobTitle: string;
  frequency: number;
  averageCandidateScore: number;
}