# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Resume Shortlisting System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Job posting operations
  - name: candidates
    description: Candidate and resume operations
  - name: analysis
    description: AI analysis and scoring operations
  - name: dashboard
    description: Dashboard summary operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job postings
      responses:
        "200":
          description: List of job postings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job posting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /jobs/{id}/candidates:
    get:
      operationId: listJobCandidates
      tags: [candidates]
      summary: List candidates for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"

  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate (submit resume text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"

  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /candidates/{id}/analyze:
    post:
      operationId: analyzeCandidate
      tags: [analysis]
      summary: Run AI analysis on a candidate for a specific job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"

  /candidates/{id}/summary:
    get:
      operationId: getCandidateSummary
      tags: [analysis]
      summary: Get AI-generated pros/cons summary for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateSummary"

  /candidates/{id}/feedback:
    get:
      operationId: getCandidateFeedback
      tags: [analysis]
      summary: Get AI-generated rejection feedback for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateFeedback"

  /candidates/{id}/blind:
    get:
      operationId: getCandidateBlind
      tags: [candidates]
      summary: Get candidate with PII hidden for blind hiring
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate with PII removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateBlind"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get overall hiring dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/top-candidates:
    get:
      operationId: getTopCandidates
      tags: [dashboard]
      summary: Get top-ranked candidates across all jobs
      responses:
        "200":
          description: Top candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateRank"

  /dashboard/skill-gaps:
    get:
      operationId: getSkillGaps
      tags: [dashboard]
      summary: Get aggregated skill gap analysis across all jobs
      responses:
        "200":
          description: Skill gap insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillGap"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Job:
      type: object
      required: [id, title, description, requiredSkills, experienceYears, status, candidateCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        status:
          type: string
          enum: [open, closed, paused]
        candidateCount:
          type: integer
        createdAt:
          type: string

    JobInput:
      type: object
      required: [title, description, requiredSkills, experienceYears]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        requiredSkills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        status:
          type: string
          enum: [open, closed, paused]

    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        status:
          type: string
          enum: [open, closed, paused]

    Candidate:
      type: object
      required: [id, name, email, resumeText, jobId, overallScore, skillSynergyScore, velocityScore, status, verifiedProofOfWork, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        resumeText:
          type: string
        jobId:
          type: integer
        overallScore:
          type: number
        skillSynergyScore:
          type: number
        velocityScore:
          type: number
        status:
          type: string
          enum: [pending, analyzed, shortlisted, rejected]
        verifiedProofOfWork:
          type: boolean
        githubUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        extractedSkills:
          type: array
          items:
            type: string
        createdAt:
          type: string

    CandidateInput:
      type: object
      required: [name, email, resumeText, jobId]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        resumeText:
          type: string
          minLength: 1
        jobId:
          type: integer
        githubUrl:
          type: string
        portfolioUrl:
          type: string
        location:
          type: string
        university:
          type: string
        yearsExperience:
          type: integer

    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, analyzed, shortlisted, rejected]
        githubUrl:
          type: string
        portfolioUrl:
          type: string

    CandidateBlind:
      type: object
      required: [id, resumeText, jobId, overallScore, skillSynergyScore, velocityScore, status, verifiedProofOfWork, extractedSkills]
      properties:
        id:
          type: integer
        resumeText:
          type: string
        jobId:
          type: integer
        overallScore:
          type: number
        skillSynergyScore:
          type: number
        velocityScore:
          type: number
        status:
          type: string
          enum: [pending, analyzed, shortlisted, rejected]
        verifiedProofOfWork:
          type: boolean
        extractedSkills:
          type: array
          items:
            type: string

    AnalyzeRequest:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer

    AnalysisResult:
      type: object
      required: [candidateId, overallScore, skillSynergyScore, velocityScore, matchedSkills, missingSkills, summary, recommendation]
      properties:
        candidateId:
          type: integer
        overallScore:
          type: number
        skillSynergyScore:
          type: number
        velocityScore:
          type: number
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        summary:
          type: string
        recommendation:
          type: string
          enum: [shortlist, reject, hold]

    CandidateFeedback:
      type: object
      required: [candidateId, emailSubject, emailBody, improvementAreas]
      properties:
        candidateId:
          type: integer
        emailSubject:
          type: string
        emailBody:
          type: string
        improvementAreas:
          type: array
          items:
            type: string

    DashboardSummary:
      type: object
      required: [totalJobs, totalCandidates, shortlisted, rejected, pending, averageScore, topScoringJob]
      properties:
        totalJobs:
          type: integer
        totalCandidates:
          type: integer
        shortlisted:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
        averageScore:
          type: number
        topScoringJob:
          type: ["string", "null"]

    CandidateRank:
      type: object
      required: [candidateId, candidateName, jobTitle, overallScore, skillSynergyScore, velocityScore, status]
      properties:
        candidateId:
          type: integer
        candidateName:
          type: string
        jobTitle:
          type: string
        overallScore:
          type: number
        skillSynergyScore:
          type: number
        velocityScore:
          type: number
        status:
          type: string

    CandidateSummary:
      type: object
      required: [candidateId, overallScore, skillSynergyScore, velocityScore, verdict, pros, cons, highlights, riskFlags, skillsFound, skillsMissing, verifiedProofOfWork]
      properties:
        candidateId:
          type: integer
        overallScore:
          type: number
        skillSynergyScore:
          type: number
        velocityScore:
          type: number
        verdict:
          type: string
          enum: [shortlist, hold, reject]
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        riskFlags:
          type: array
          items:
            type: string
        skillsFound:
          type: array
          items:
            type: string
        skillsMissing:
          type: array
          items:
            type: string
        verifiedProofOfWork:
          type: boolean

    SkillGap:
      type: object
      required: [skill, jobTitle, frequency, averageCandidateScore]
      properties:
        skill:
          type: string
        jobTitle:
          type: string
        frequency:
          type: integer
        averageCandidateScore:
          type: number