# lib/api-zod/src/generated/types/candidateSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Resume Shortlisting System API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateSummaryVerdict } from './candidateSummaryVerdict';

export interface CandidateSummary {
  candidateId: number;
  overallScore: number;
  skillSynergyScore: number;
  velocityScore: number;
  verdict: CandidateSummaryVerdict;
  pros: string[];
  cons: string[];
  highlights: string[];
  riskFlags: string[];
  skillsFound: string[];
  skillsMissing: string[];
  verifiedProofOfWork: boolean;
}